==> 1_env.qmd <==
```{r setupenv, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(sf)
ggplot2::theme_set(bayesplot::theme_default())
```

# Environment

```{r envdata, eval=F}
# data
paracou <- vroom::vroom("data/paracou.tsv") %>% 
  filter(CensusYear == 2019)
paracouXY <- paracou %>% 
  filter(!is.na(Xutm), !is.na(Yutm)) %>% 
  st_as_sf(coords = c("Xutm", "Yutm"),
           crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')

# rasters
paracouXY$TWI <- raster::extract(raster::raster("data/TWI_1m.tif"), paracouXY)
paracouXY <- mutate(paracouXY, TWI = ifelse(TWI < 0, 0, TWI))
paracouXY$HAND <- raster::extract(raster::raster("data/RelativeElevation_1m.tif"), paracouXY)
paracou <- paracou %>% 
  left_join(select(paracouXY, idTree, TWI, HAND) %>% st_drop_geometry())
  

# shapefiles
readshp <- function(path){
  shp <- st_read(path, crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')
  shp$geometry <- lwgeom::lwgeom_make_valid(st_as_sfc(shp))
  return(shp)
}

paracou <- paracou %>% 
  left_join(st_intersection(readshp("data/Hydromorphy/Hydromorphy.shp"), paracouXY) %>% 
              st_drop_geometry() %>% 
              select(idTree, Hydro)) %>% 
  left_join(st_intersection(readshp("data/WaterLogging/Waterlogging.shp"), paracouXY) %>% 
              st_drop_geometry() %>% 
              select(idTree, LevelWater)) %>% 
  left_join(st_intersection(readshp("data/WaterTable/WaterTable.shp"), paracouXY) %>% 
              st_drop_geometry() %>% 
              select(idTree, TypeWatEN)) %>% 
  left_join(st_intersection(readshp("data/Drainages/Drainages.shp"), paracouXY) %>% 
              st_drop_geometry() %>% 
              select(idTree, TypDrainEN))

# save
vroom::vroom_write(paracou, "data/paracou_env.tsv")

# ind env
inds <- readxl::read_xlsx("data/TREE_list_Paracou.xlsx") %>% 
  filter(!is.na(XUTM), !is.na(YUTM))
inds_xy <- inds %>% 
  st_as_sf(coords = c("XUTM", "YUTM"),
           crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')
inds$hand <- extract(rast("data/RelativeElevation_1m.tif"), vect(inds_xy))$RelativeElevation
vroom::vroom_write(inds, "data/inds_env.tsv")

# ind env all
bind_rows(
  inds %>% 
  separate("Species...1", c("Genus", "Species")) %>% 
  select(Genus, Species, Plot, hand) %>% 
  mutate(sampled = 1) %>% 
  mutate(Plot = as.character(Plot)) %>% 
    filter(paste(Genus, Species) %in% 
             paste(paracou$Genus, paracou$Species)),
  paracou %>% 
  select(Genus, Species, Plot, HAND) %>% 
  rename(hand = HAND) %>% 
  mutate(sampled = 0)
) %>% 
  vroom::vroom_write("data/inds_env2.tsv")
```

## TWI: Topographic Wetness Index

```{r twi}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  group_by(Species) %>% 
  ggplot(aes(x = species, y = log(TWI+1))) +
  geom_boxplot() +
  coord_flip()
```

```{r}
vroom::vroom("data/paracou_env.tsv") %>% 
  group_by(Genus, Species) %>% 
  summarise(twi_min = min(TWI),
            twi_q1 = quantile(TWI, 0.01),
            twi_median = median(TWI),
            twi_mean = mean(TWI),
            twi_q99 = quantile(TWI, 0.99),
            twi_max = max(TWI)) %>% 
  write_tsv("twi_quantiles.tsv")
```

```{r}
read_tsv("twi_quantiles.tsv") %>% 
  gather(quantile, value, -Genus, -Species) %>% 
  mutate(quantile = gsub("twi_", "", quantile)) %>% 
  ggplot(aes(paste(Genus, Species), value, col = quantile)) +
  geom_point() +
  theme_bw() +
  coord_flip() +
  theme(axis.text.y = element_text(face = "italic"),
        axis.title = element_blank())
```

## HAND: Height Above Nearest Drainage

```{r hand}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  ggplot(aes(x = species, y = log(HAND+1))) +
  geom_boxplot() +
  coord_flip()
```

```{r}
dat <- vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  select(species, HAND) %>% 
  group_by(species) %>% 
  bind_rows(data.frame(species = "All",
                       HAND = terra::rast("data/RelativeElevation_1m.tif") %>%
                         as.vector() %>% na.omit() %>% sample(10^4)))
g <- ggplot(dat, aes(HAND)) +
  geom_histogram() +
  facet_wrap(~ species, scales = "free_y") +
  geom_vline(aes(xintercept = HAND, col = "Maximum"),
             data = group_by(dat, species) %>% 
               summarise(HAND = max(HAND))) +
  geom_vline(aes(xintercept = HAND, col ="99th quantile"),
             data = group_by(dat, species) %>% 
               summarise(HAND = quantile(HAND, .99))) +
  scale_y_sqrt() +
  theme(strip.text.x = element_text(size = 8), legend.position = "bottom")
ggsave("hand.png", g, bg = "white", width = 14)
```

```{r}
plot(terra::rast("data/RelativeElevation_1m.tif") > 30)
```

```{r}
dat <- vroom::vroom("data/inds_env2.tsv")
ggplot(dat, aes(paste(Genus, Species), hand)) +
  geom_violin(data = filter(dat, sampled == 0)) +
  geom_point(data = filter(dat, sampled == 1),
             col = "#93c4dd", size = 1) +
  geom_point(data = group_by(dat, Genus, Species) %>% 
                 summarise(hand = median(hand)),
             col = "#93c4dd", size = 3) +
  geom_point(data = group_by(dat, Genus, Species) %>% 
                 summarise(hand = quantile(hand, .99)),
             col = "#e0697c", size = 3) +
  theme_bw() +
  xlab("") + ylab("HAND (m)") +
  theme(axis.text.y = element_text(face = "italic")) +
  coord_flip()
```

## Hydromorphy

```{r hydro}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  filter(!is.na(Hydro)) %>% 
  ggplot(aes(x = species, fill = as.factor(Hydro))) +
  geom_bar(position = "dodge") +
  coord_flip() +
  scale_fill_discrete("Hydromorphy")
```

## Water logging

```{r wl}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  filter(!is.na(LevelWater)) %>% 
  ggplot(aes(x = species, fill = as.factor(LevelWater))) +
  geom_bar(position = "dodge") +
  coord_flip() +
  scale_fill_discrete("Water\nlogging")
```

## Water table depth

```{r wtd}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  filter(!is.na(TypeWatEN)) %>% 
  ggplot(aes(x = species, fill = as.factor(TypeWatEN))) +
  geom_bar(position = "dodge") +
  coord_flip() +
  scale_fill_discrete("") +
  guides(fill=guide_legend(nrow=3)) +
  theme(legend.position = "bottom")
```

## Drainage type

```{r drain}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(species =  paste(Genus, Species)) %>% 
  filter(!is.na(TypDrainEN)) %>% 
  ggplot(aes(x = species, fill = as.factor(TypDrainEN))) +
  geom_bar(position = "dodge") +
  coord_flip() +
  scale_fill_discrete("Drainage")
```

## Correlations

```{r corr, fig.height=10, fig.width=10}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  mutate(TWI = log(TWI+1), HAND = log(HAND + 1)) %>% 
  select(TWI, HAND, Hydro, LevelWater, TypeWatEN, TypDrainEN) %>% 
  na.omit() %>% 
  fastDummies::dummy_cols(c("Hydro", "LevelWater", "TypeWatEN", "TypDrainEN"), remove_selected_columns = T) %>% 
  cor() %>% 
  corrplot::corrplot(type = "lower", method = "number", tl.cex = 0.8)
```

## Species data

```{r spenvdata}
#| message: false
vroom::vroom("data/paracou_env.tsv") %>% 
  group_by(Genus, Species) %>% 
  summarise(TWI = mean(log(TWI+1), na.omit = T), 
            HAND = mean(log(HAND+1), na.omit = T)) %>% 
  vroom::vroom_write("data/species_env")
```

```{r spenvdatatab}
#| message: false
vroom::vroom("data/species_env") %>% 
  kable(caption = "Envrionmental descriptor per species.")
```
