==> 3_tests.qmd <==
```{r setuptests, include=FALSE, eval=T}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(tidyverse)
library(vroom)
library(sf)
library(patchwork)
library(tidyverse)
library(readxl)
ggplot2::theme_set(bayesplot::theme_default())
```

# Tests

## Figure with error

```{r data}
#| message: false
#| warning: false
data <- read_excel("data/camille/DATABASE_Canopy.xlsx", "All_data_Ind") %>% 
  separate("Species...1", c("Genus", "Species")) %>% 
  mutate(SSMleaf = Ptlp_wet - P50leaf, 
         SSMstem = Ptlp_wet - P50stem) %>% 
  select(Genus, Species, Ptlp_wet, P50leaf, P50stem,  SSMleaf, SSMstem, SegP50,
         gmin_leaf, gmin_bark, `Al:As`, SegLSC_leaf, Pmd_18,
         PLCleaf_18, PLCstem_18, Mortality) %>% 
  rename(Ptlp = Ptlp_wet, gmin = gmin_leaf, gbark = gmin_bark,
         SegLSC = SegLSC_leaf, Pmd = Pmd_18, PLCleaf = PLCleaf_18,
         PLCstem = PLCstem_18, AlAs = `Al:As`) %>% 
  gather(variable, value, -Genus, -Species) %>% 
  group_by(Genus, Species, variable) %>% 
  summarise(median = median(value, na.rm = T), sd  = sd(value, na.rm = T)) %>% 
  mutate(sd = ifelse(is.na(sd), 0, sd)) %>% 
  gather(metric, value, -Genus, -Species, -variable) %>% 
  mutate(variable = paste0(variable, "_", metric)) %>% 
  select(-metric) %>% 
  pivot_wider(values_from = value, names_from = variable)
```

```{r p50leafstemv1, fig.width=6, fig.height=6}
#| message: false
#| warning: false
ggplot(data, aes(P50stem_median, P50leaf_median)) + 
  geom_smooth(method = lm, color = "darkred", fill = "#000000", se = T) +
  geom_abline(linetype = "dashed", size = 0.8) +
  geom_point(size = 4, color = "black") +
  geom_errorbar(aes(ymin = P50leaf_median - P50leaf_sd, 
                    ymax = P50leaf_median + P50leaf_sd)) +
  geom_errorbar(aes(xmin = P50stem_median - P50stem_sd, 
                    xmax = P50stem_median + P50stem_sd)) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 24), 
        axis.text=element_text(color = "black"), 
        legend.position = "none") +
  scale_x_continuous(limits = c(-8.5,-1), breaks = seq(-8,-1,1)) +
  scale_y_continuous(limits = c(-7,-1), breaks = seq(-5,-2,1)) +
  ylab(expression(P[paste(50,",",leaf)]~(MPa))) + 
  xlab (expression(P[paste(50,",",stem)]~(MPa)))
```

```{r p50leafstemv2, fig.width=6, fig.height=6}
#| message: false
#| warning: false
ggplot(data, aes(P50stem_median, P50leaf_median)) + 
  geom_smooth(method = lm, color = "darkred", fill = "#000000", se = T) +
  geom_abline(linetype = "dashed", size = 0.8) +
  geom_point(size = 4, color = "black") +
  geom_errorbar(aes(ymin = P50leaf_median - P50leaf_sd,
                    ymax = P50leaf_median + P50leaf_sd,
                    width = 0.2)) +
  geom_errorbar(aes(xmin = P50stem_median - P50stem_sd,
                    xmax = P50stem_median + P50stem_sd,
                    width = 0.2)) +
  theme_bw() +
  theme(panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        text = element_text(size = 24), 
        axis.text=element_text(color = "black"), 
        legend.position = "none") +
  scale_x_continuous(limits = c(-8.5,-1), breaks = seq(-8,-1,1)) +
  scale_y_continuous(limits = c(-7,-1), breaks = seq(-5,-2,1)) +
  ylab(expression(P[paste(50,",",leaf)]~(MPa))) + 
  xlab (expression(P[paste(50,",",stem)]~(MPa)))
```

## Old

### Data

```{r dataold}
#| message: false
#| warning: false
ind <- readxl::read_excel("data/camille/DATABASE_Canopy.xlsx", "All_data_Ind") %>% 
  separate("Species...1", c("Genus", "Species")) %>% 
  filter(!is.na(XUTM)) %>% 
  mutate(SSMleaf = Ptlp_wet - P50leaf, 
         SSMstem = Ptlp_wet - P50stem) %>% 
  select(Genus, Species, Ptlp_wet, P50leaf, P50stem,  SSMleaf, SSMstem, SegP50,
         gmin_leaf, gmin_bark, `Al:As`, SegLSC_leaf, Pmd_18,
         PLCleaf_18, PLCstem_18, Mortality) %>% 
  rename(Ptlp = Ptlp_wet, gmin = gmin_leaf, gbark = gmin_bark,
         SegLSC = SegLSC_leaf, Pmd = Pmd_18, PLCleaf = PLCleaf_18,
         PLCstem = PLCstem_18)
indxy <- readxl::read_excel("data/camille/DATABASE_Canopy.xlsx", "All_data_Ind") %>% 
  filter(!is.na(XUTM)) %>% 
  st_as_sf(coords = c("XUTM", "YUTM"),
           crs = '+proj=utm +zone=22 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0')
ind$HAND <- raster::extract(raster::raster("data/RelativeElevation_1m.tif"), indxy)
rm(indxy)
agr <- readxl::read_excel("data/camille/DATABASE_Canopy.xlsx", "All_data_Ind") %>% 
  filter(!is.na(XUTM)) %>% 
  select("Species...1") %>% 
  separate("Species...1", c("Genus", "Species")) %>% 
  left_join(vroom("data/species_growth.tsv"))
ind$AGR <- agr$Plateau
```

### Normality

```{r distributionsraw}
#| message: false
#| warning: false
ind %>% 
  gather(variable, value, -Genus, -Species) %>% 
  na.omit() %>% 
  ggplot(aes(value)) +
  geom_histogram() +
  facet_wrap(~ variable, scales = "free") +
  ggtitle("Distribution raw")
```

```{r distributiontransformed}
#| message: false
#| warning: false
ind %>% 
  gather(variable, value, -Genus, -Species) %>% 
  na.omit() %>% 
  ggplot(aes(log(abs(value)))) +
  geom_histogram() +
  facet_wrap(~ variable, scales = "free") +
  ggtitle("Distribution log and abs transformed")
```

### Figure 1

```{r}
#| message: false
#| warning: false
m1 <- lm(P50stem ~ P50leaf, ind)
m2 <- lm(gbark ~ gmin, ind)
m3 <- lm(gbark ~ P50stem, ind)
m4 <- lm(gmin ~ P50leaf, ind)
sjPlot::tab_model(m1, m2, m3, m4)
```

```{r fig1, fig.height=6, fig.width=6}
#| message: false
#| warning: false
g1 <- ggplot(ind, aes(P50stem, P50leaf)) +
  geom_smooth(method = "lm", formula = y ~ x, col = "red") +
  geom_abline(linetype = "dashed") +
  geom_point() +
  xlab(expression(P[50~stem]~(MPa))) +
  ylab(expression(P[50~leaf]~(MPa))) +
  ggtitle(paste0("R²=", round(broom::glance(m1)$r.squared, 3),
          ", p=", round(broom::glance(m1)$p.value, 3)))
g2 <- ggplot(ind, aes(gbark, gmin)) +
  geom_abline(linetype = "dashed") +
  geom_point() +
  ylab(expression(g[min]~(mmol~m^{-2}~s^{-1}))) + ylim(0, 10) +
  xlab(expression(g[bark]~(mmol~m^{-2}~s^{-1}))) +
  ggtitle(paste0("R²=", round(broom::glance(m2)$r.squared, 3),
          ", p=", round(broom::glance(m2)$p.value, 3)))
g3 <- ggplot(ind, aes(P50stem, gbark)) +
  geom_point() +
  xlab(expression(P[50~stem]~(MPa))) +
  ylab(expression(g[bark]~(mmol~m^{-2}~s^{-1}))) +
  ggtitle(paste0("R²=", round(broom::glance(m3)$r.squared, 3),
          ", p=", round(broom::glance(m3)$p.value, 3)))
g4 <- ggplot(ind, aes(P50leaf, gmin)) +
  geom_point() +
  ylab(expression(g[min]~(mmol~m^{-2}~s^{-1}))) + ylim(0, 10) +
  xlab(expression(P[50~leaf]~(MPa))) +
  ggtitle(paste0("R²=", round(broom::glance(m4)$r.squared, 3),
          ", p=", round(broom::glance(m4)$p.value, 3)))
(g1 + g2) / (g3 + g4)
```

### Figure 2

```{r fig2, fig.height=7, fig.width=7}
#| message: false
#| warning: false
g1 <- ind %>% 
  select(SSMleaf, SSMstem, SegP50) %>% 
  gather(variable, value, -SegP50) %>% 
  na.omit() %>% 
  ggplot(aes(SegP50, value, col = variable)) +
  geom_smooth(method = "lm") +
  geom_point() +
  scale_color_manual(guide = "none", values = c("darkgreen", "orange")) +
  ggpubr::stat_cor(aes(label =  paste(..p.label..)), label.y.npc = 0.7) +
  ggpubr::stat_regline_equation(aes(label =  paste(..adj.rr.label..))) +
  xlab(expression(Seg[P50]~(MPa))) +
  ylab(expression(SSM[leaf]~or~SSM[stem]~(MPa)))
g2 <- ggplot(ind, aes(SegLSC, gmin)) +
  geom_point() +
  ggpubr::stat_cor(aes(label =  paste(..p.label..)), label.y.npc = 0.8) +
  ggpubr::stat_regline_equation(aes(label =  paste(..adj.rr.label..))) +
  ylab(expression(g[min]~(mmol~m^{-2}~s^{-1}))) + ylim(0, 10) +
  xlab(expression(Seg[LSC]~(10^{5}~kg~m^{-1}~Mpa^{-1}~s^{-1})))
g3 <- ind %>% 
  select(gmin, gbark, `Al:As`) %>% 
  gather(variable, value, -`Al:As`) %>% 
  na.omit() %>% 
  ggplot(aes(value, `Al:As`, col = variable)) +
  geom_smooth(method = "lm") +
  geom_point() +
  scale_color_manual(guide = "none", values = c("darkgreen", "orange")) +
  ggpubr::stat_cor(aes(label =  paste(..p.label..)), label.y.npc = 0.7) +
  ggpubr::stat_regline_equation(aes(label =  paste(..adj.rr.label..))) +
  ylab("Al:As") +
  xlab(expression(g[min]~or~g[bark]~(mmol~m^{-2}~s^{-1})))
g4 <- ind %>% 
  select(P50leaf, P50stem, `Al:As`) %>% 
  gather(variable, value, -`Al:As`) %>% 
  na.omit() %>% 
  ggplot(aes(value, `Al:As`, col = variable)) +
  geom_point() +
  scale_color_manual(guide = "none", values = c("darkgreen", "orange")) +
  ggpubr::stat_cor(aes(label =  paste(..p.label..)), label.y.npc = 0.7) +
  ggpubr::stat_regline_equation(aes(label =  paste(..adj.rr.label..))) +
  ylab("Al:As") +
  xlab(expression(P[50~leaf]~or~P[50~stem]~(MPa)))
(g1 + g2) / (g3 + g4)
```

### Figure 3

```{r}
#| message: false
#| warning: false
data <- select(ind, PLCstem, P50stem, 
               SSMstem, SegP50) %>% 
  na.omit()
m0 <- lm(log(PLCstem) ~ P50stem, data)
m1 <- lm(log(PLCstem) ~ SSMstem, data)
m2 <- lm(log(PLCstem) ~ SegP50, data)
anova(m0, m1, m2)
```

### Figure 5

```{r}
#| message: false
#| warning: false
data <- ind %>% 
  rename(AlAs = `Al:As`) %>% 
  select(HAND, Ptlp, P50leaf, P50stem, gmin, gbark, AlAs) %>% 
  na.omit()
lm(HAND ~ Ptlp + P50leaf + P50stem + log(gmin) + log(gbark) + log(AlAs), data) %>% 
  step(trace = F) %>% 
  sjPlot::tab_model()
```

```{r fig5}
#| message: false
#| warning: false
datag <- data %>% 
  mutate(gmin = log(gmin), gbark = log(gbark), AlAs = log(AlAs)) %>% 
  gather(variable, value, -HAND) %>% 
  na.omit() %>% 
  mutate(variable_long = recode(variable,
                                "AlAs" = "log(Al:As)",
                                "Ptlp" = "P[tlp]~(MPa)",
                                "P50leaf" = "P[50~leaf]~(MPa)",
                                "P50stem" = "P[50~stem]~(MPa)",
                                "gmin" = "log(g[min])~(mmol~m^{-2}~s^{-1})",
                                "gbark" = "log(g[bark])~(mmol~m^{-2}~s^{-1})"
  ))
ggplot(datag, aes(HAND, value)) +
  geom_point() +
  facet_wrap(~ variable_long, scales = "free", labeller = label_parsed) +
  ylab("") + xlab("HAND (m)")
```

### Figure 6

```{r}
#| message: false
#| warning: false
data <- ind %>% 
  rename(AlAs = `Al:As`) %>% 
  select(Mortality, Ptlp, P50leaf, P50stem, gmin, gbark, AlAs) %>% 
  na.omit()
lm(Mortality ~ Ptlp + P50leaf + P50stem + log(gmin) + log(gbark) + log(AlAs), data) %>% 
  step(trace = F) %>% 
  sjPlot::tab_model()
```

```{r fig6}
#| message: false
#| warning: false
datag <- data %>% 
  mutate(gmin = log(gmin), gbark = log(gbark), AlAs = log(AlAs)) %>% 
  gather(variable, value, -Mortality) %>% 
  na.omit() %>% 
  mutate(variable_long = recode(variable,
                                "AlAs" = "log(Al:As)",
                                "Ptlp" = "P[tlp]~(MPa)",
                                "P50leaf" = "P[50~leaf]~(MPa)",
                                "P50stem" = "P[50~stem]~(MPa)",
                                "gmin" = "log(g[min])~(mmol~m^{-2}~s^{-1})",
                                "gbark" = "log(g[bark])~(mmol~m^{-2}~s^{-1})"
  ))
ggplot(datag, aes(value, Mortality)) +
  geom_point() +
  facet_wrap(~ variable_long, scales = "free", labeller = label_parsed) +
  ylab(expression(Mortality~("%"~year^{-1}))) + xlab("") +
   geom_smooth(method = "lm", data = filter(datag, variable %in% c("gbark")))
```

```{r}
#| message: false
#| warning: false
m1 <- lm(Mortality ~ log(gbark), ind)
sjPlot::tab_model(m1)
```

```{r}
#| message: false
#| warning: false
ggplot(data, aes(Mortality, gbark)) +
  geom_smooth(method = "lm", formula = y ~ log(x), col = "red") +
  geom_point() +
  ylab(expression(Mortality~("%"~year^{-1}))) +
  xlab(expression(g[bark]~(mmol~m^{-2}~s^{-1}))) +
  ggtitle(paste0("R²=", round(broom::glance(m1)$r.squared, 3),
          ", p=", round(broom::glance(m1)$p.value, 3)))
```

### Figure 7

```{r }
#| message: false
#| warning: false
data <- ind %>% 
  rename(AlAs = `Al:As`) %>% 
  select(AGR, Ptlp, P50leaf, P50stem, gmin, gbark, AlAs) %>% 
  na.omit()
lm(log(AGR) ~ Ptlp + P50leaf + P50stem + log(gmin) + log(gbark) + log(AlAs), data) %>% 
  step(trace = F) %>% 
  sjPlot::tab_model()
```

```{r fig7}
#| message: false
#| warning: false
datag <- data %>% 
  mutate(gmin = log(gmin), gbark = log(gbark), AlAs = log(AlAs)) %>% 
  gather(variable, value, -AGR) %>% 
  na.omit() %>% 
  mutate(variable_long = recode(variable,
                                "AlAs" = "log(Al:As)",
                                "Ptlp" = "P[tlp]~(MPa)",
                                "P50leaf" = "P[50~leaf]~(MPa)",
                                "P50stem" = "P[50~stem]~(MPa)",
                                "gmin" = "log(g[min])~(mmol~m^{-2}~s^{-1})",
                                "gbark" = "log(g[bark])~(mmol~m^{-2}~s^{-1})"
  ))
ggplot(datag, aes(value, AGR)) +
  geom_point() +
  facet_wrap(~ variable_long, scales = "free", labeller = label_parsed) +
  ylab(expression(Annual~growth~rate~("cm"~year^{-1}))) + xlab("") +
  scale_y_log10() +
  geom_smooth(method = "lm", data = filter(datag, variable %in% c("gbark", "gmin", "P50stem")))
```

```{r, eval=F}
#| message: false
#| warning: false
datag2 <- select(data, gbark, gmin, P50stem, AGR) %>% 
  gather(variable, value, -AGR) %>% 
  na.omit() %>% 
  mutate(variable_long = recode(variable,
                                "P50stem" = "P[50~stem]~(MPa)~~p==0.043",
                                "gmin" = "log(g[min])~(mmol~m^{-2}~s^{-1})~~p==0.006",
                                "gbark" = "log(g[bark])~(mmol~m^{-2}~s^{-1})~~p==0.003"
  ))

ggplot(datag2, aes(value, AGR)) +
  geom_point() +
  scale_y_log10() +
  ylab(expression(Annual~growth~rate~("cm"~year^{-1}))) + xlab("") +
  ggtitle("log(AGR) ~ P50 + log(gmin) + log(gbark)", "N=22, R²=0.467") +
  facet_wrap(~ variable_long, scales = "free", labeller = label_parsed) +
  geom_smooth(method = "lm", formula = y ~ log(x), col = "red",
              data = filter(datag2, variable %in% c("gmin", "gbark"))) +
  geom_smooth(method = "lm", formula = y ~ x, col = "red",
              data = filter(datag2, variable %in% c("P50stem")))
```

### Figure S2

```{r figs2}
#| message: false
#| warning: false
ind %>% 
  select(-Genus, -Species) %>% 
  cor(use = "pairwise.complete.obs") %>% 
  corrplot::corrplot(type = "upper",
                     method = "ellipse",
                     col = colorRampPalette(c("blue", "white","red"))(100))
```

### Figure S4

```{r figs4}
#| message: false
#| warning: false
ind %>% 
  select(Genus, Species, HAND) %>% 
  group_by(Genus, Species) %>% 
  mutate(Hmed = median(HAND)) %>% 
  arrange(Hmed) %>% 
  mutate(type = "Terra firme") %>% 
  mutate(type = ifelse(Species %in% c("bidentata", "rosea", "globulifera", "melinonii"), 
                       "Generalist", type)) %>% 
  mutate(type = ifelse(Species %in% c("falcata", "glabra", "hexapetala", "persistens",
                                      "opacum"),
                       "Bottomland", type)) %>%
  ggplot(aes(reorder(paste(Genus, Species), Hmed), HAND, col = type)) +
  geom_boxplot() +
  coord_flip() +
  xlab("") + ylab("HAND (m)") +
  theme(axis.text.y = element_text(face = "italic")) +
  scale_color_manual(guide = "none", values = c("green", "orange", "firebrick"))
```

### ACP

```{r}
#| message: false
#| warning: false
library("FactoMineR")
library("factoextra")
library(NbClust)
library(tidyverse)
library(RColorBrewer)
library(ggsci)

Tab <- read.csv("data/ACP_canopy_strategy_traits_gapfill_19sp.csv",
                header = T, sep = ";", na.strings = "NA")

#enlever les colonnes en trop
Tab$X<-NULL
Tab$X.1<-NULL
# dim(Tab)

#PCA traits without supplementary quantitative variables
#Tab 3 only has active traits (without supp quanti)
Tab3<-Tab[1:20,1:33]
# dim(Tab3)
# head(Tab3[, 1:33], 50) #Pour loader seulement les colonnes intéressantes
Tab3<-subset(Tab,select=c(Species, SSMleaf, SSMstem, SegP50,
                          Ptlp, P50stem, P50leaf, gmin, gbark, Al.As, SegLSC))
Tab3<-na.omit(Tab3)
rownames(Tab3)<-Tab3$Species
Tab4 <- Tab3[ , - c(1)]
names(Tab4) <- c("SSM[leaf]",
                 "SSM[stem]",
                 "Seg[Psi(50)]",
                 "pi[tlp]",
                 'italic(Psi)[paste(50,",",stem)]',
                 'italic(Psi)[paste(50,",",leaf)]',
                 "italic(g)[min]",
                 "italic(g)[bark]",
                 'paste(A[l],":",A[s])',
                 "Seg[LSC]" )
res.pca <- PCA(Tab4, quanti.sup = 1:3, graph = FALSE)
g <- fviz_pca_var(res.pca, col.var = "contrib", axes = c(1,2), parse = TRUE,
                  gradient.cols = c("#FFCC00","#990099", "#CC0033"),
                  arrowsize = 1, labelsize = 6, 
                  axes.linetype = "dashed", 
                  pointsize = 3, repel = T,
                  col.ind = "black", title = "",
                  col.quanti.sup = "#6666FF",
                  invisible = "none",
                  axis.text=element_text(size=16),
                  ggtheme=theme(axis.text=element_text(size=22), 
                                axis.title=element_text(size=22),
                                legend.title = element_text(size = 14),
                                legend.text = element_text(size = 14),
                                panel.background = element_rect(fill = "white", colour = "black",
                                                                size = 1, linetype = "solid"),
                                panel.grid.major = element_line(size = 0.5, linetype = 'solid',
                                                                colour = "white"),
                                panel.grid.minor = element_line(size = 0.25, linetype = 'solid',
                                                                colour = "white")))

g$layers[[5]]$geom_params$parse <- TRUE
g
```
